Update README.md
lighter version
The error SwissEph file 'seas_18.se1' not found happens because calculating the positions of asteroids like Chiron, Ceres, Pallas, Juno, and Vesta requires a separate, large ephemeris file that is not included with the pyswisseph library by default.

To fix this and remove the dependency on users installing external files, we can remove the calculations for these specific minor bodies. The main planets (Sun-Pluto) and the Lunar Nodes do not require these external files and can be calculated using the built-in capabilities of pyswisseph.

Here is the complete, corrected code with those dependencies removed. This version should run without any file-not-found errors.